fix: 이력서 다운로드 배치 잘못 구성되어 재구성 및 상태 변경 모달 초기작업
