+++ src/app/applications/[id]/components/modal/EditStatusForm.tsx
@@ -1,0 +1,7 @@
+export default function EditStatusForm({
+  handleCloseModal,
+}: {
+  handleCloseModal: () => void;
+}) {
+  return <form action=''></form>;
+}
